- Implementado tratamento de exceções - Ajustes no README - Corrigidos bugs

+++ src/main/java/com/alves/vitor/DigitalAccounts/application/usecases/account/DeleteAccount.java
@@ -2,6 +2,7 @@ package com.alves.vitor.DigitalAccounts.application.usecases.account;
 
 import com.alves.vitor.DigitalAccounts.application.gateways.AccountRepository;
 import com.alves.vitor.DigitalAccounts.domain.entity.Account;
+import com.alves.vitor.DigitalAccounts.domain.exceptions.AccountNotFoundException;
 
 public class DeleteAccount {
     private final AccountRepository repository;
@@ -11,6 +12,12 @@ public class DeleteAccount {
     }
 
     public Account delete(String agency, String number) {
+        Account exists = repository.findByAgencyAndNumber(agency, number);
+
+        if (exists == null) {
+            throw new AccountNotFoundException();
+        }
+
         return repository.delete(agency, number);
     }
 }
